Refactor xtask commands into separate files
