+++ dev-tools/xtask/src/verify_libraries.rs
@@ -83,7 +83,7 @@ fn verify_executable(
 
     Ok(())
 }
-pub fn cmd_verify_libraries() -> Result<()> {
+pub fn run_cmd() -> Result<()> {
     let metadata = load_workspace()?;
     let mut config_path = metadata.workspace_root;
     config_path.push(".cargo/xtask.toml");
